Issue #125 - basic constraint-only profile support
